Add all remaining classes of the EN SDK
This commit adds all the remaining classes from the EN SDK to the public fork.

It also adds several annotations to help keep the fork as close to the original
as possible, which unfortunately requires the use of Java 9 to compile the
project. This means that, to compile from Android Studio, the JDK location must
be manually set to point to Java 9.

Change-Id: I4879cfe812b2c9179e47ab62a0c537929a3245b8